Update recording method in navirice_record.
Now you can have 3 buttons. The first one is just like it was
previously.
The next two run haar-cascades on the images, and depending on the mode
will only save if there is or is not a head detected. This will help us
avoid saving images of face that the haarcascades don't catch, as well
as avoid saving images of nonfaces that the haarcascades catch by
accident.